broke out search and pagesize to separate toolbar
